Refactor: extract ValidationSession from entry_point.dart
entry_point.dart had grown to ~900 lines, mixing CLI argument parsing,
config loading, and the per-skill validation workflow. The two processors
(_processSkillPaths and _processSkillDirectories) duplicated significant
per-root setup, ignore loading, and stale-ignore reporting.

Extract a new ValidationSession class that owns per-invocation state
(config, resolved rules, customRules, flags, anyFailed/anySkillsValidated)
and exposes processIndividualSkill / processSkillRoot / reportNoSkillsValidated.
entry_point.dart drops to ~360 lines and becomes a thin orchestrator.

Behavior preserved exactly:
- Original fast-fail semantics: missing-directory errors contribute to
  anyFailed but don't trigger fast-fail in their own iteration; only
  validation failures do. The session methods return bool keepGoing to
  signal this to the caller.
- Public test-visible constants (skillIsValidMsg, evaluatingDirMsg, etc.)
  moved to validation_session.dart and re-exported from entry_point.dart
  via `export ... show` so test imports keep working unchanged.
- defaultIgnoreFileName loses its @visibleForTesting annotation: it's
  used in the --generate-baseline help text in production code, so the
  original annotation was a mis-application that the move exposed.

Verified:
- dart analyze --fatal-infos reports no issues
- dart test passes all 110 tests, including the fast-fail integration test
  and all fixer tests
- dart run bin/cli.dart --skills-directory ../../skills validates all 10
  Flutter skills cleanly
